+++ pybop/_result.py
@@ -1,11 +1,6 @@
-from typing import TYPE_CHECKING
-
 import numpy as np
 
-if TYPE_CHECKING:
-    from pybop import BaseOptimiser
 from pybop import Logger, Problem, plot
-from pybop.parameters.multivariate_parameters import MultivariateParameters
 
 
 class Result:
@@ -319,108 +314,3 @@ class Result:
             Valid Plotly layout keys and their values.
         """
         return plot.contour(call_object=self, **kwargs)
-
-
-class OptimisationResult(Result):
-    """
-    Stores the result of the optimisation.
-
-    Attributes
-    ----------
-    optim : pybop.BaseOptimiser
-        The optimisation object used to generate the results.
-    time : float
-        The time taken.
-    method_name : str
-        The name of the optimiser.
-    message : str
-        The reason for stopping given by the optimiser.
-    scipy_result : scipy.optimize.OptimizeResult, optional
-        The result obtained from a SciPy optimiser.
-    """
-
-    def __init__(
-        self,
-        optim: "BaseOptimiser",
-        time: float,
-        method_name: str | None = None,
-        message: str | None = None,
-        scipy_result=None,
-    ):
-        super().__init__(
-            problem=optim.problem,
-            logger=optim.logger,
-            time=time,
-            method_name=method_name,
-            message=message,
-            scipy_result=scipy_result,
-        )
-
-
-class BayesianOptimisationResult(OptimisationResult):
-    """
-    Stores the result of a Bayesian optimisation or a Bayesian model
-    selection.
-
-    Attributes
-    ----------
-    problem: pybop.Problem
-        The optimisation problem used to generate the results.
-    x : ndarray
-        The solution of the optimisation (in model space).
-    best_cost : float
-        The cost associated with the solution x.
-    n_iterations : int or dict
-        Number of iterations performed by the optimiser. Since Bayesian
-        optimisers tend to have layers of various optimisation
-        algorithms, their iteration counts may be put individually.
-    n_evaluations : int or dict
-        Number of evaluations performed by the optimiser. Since Bayesian
-        optimisers tend to have layers of various optimisation
-        algorithms, their evaluation counts my be put individually.
-    message : str
-        The reason for stopping given by the optimiser.
-    lower_bounds: ndarray
-        The lower confidence parameter boundaries.
-    upper_bounds: ndarray
-        The upper confidence parameter boundaries.
-    posterior : MultivariateParameters
-        The probability distribution of the optimisation.
-    maximum_a_posteriori : Inputs or ndarray
-        Complementing the best observed value in `x`, this is the
-        prediction for the best parameter value.
-    log_evidence_mean : float
-        The logarithm of the evidence of the parameterization. Higher
-        values are better. May only be interpreted relative to a
-        calibration case, e.g., a test-run with synthetic data.
-    log_evidence_variance : float
-        The logarithm of the variance in the calculation of the
-        evidence. For reliable comparisons based on the evidence, should
-        be at or below the scale of the evidence itself.
-    """
-
-    def __init__(
-        self,
-        optim: "BaseOptimiser",
-        time: float | dict,
-        method_name: str | None = None,
-        message: str | None = None,
-        lower_bounds: np.ndarray | None = None,
-        upper_bounds: np.ndarray | None = None,
-        posterior: MultivariateParameters | None = None,
-        maximum_a_posteriori: np.ndarray | None = None,
-        log_evidence_mean: float | None = None,
-        log_evidence_variance: float | None = None,
-    ):
-        super().__init__(
-            optim=optim,
-            time=time,
-            method_name=method_name,
-            message=message,
-        )
-        self.lower_bounds = lower_bounds
-        self.upper_bounds = upper_bounds
-        self.posterior = posterior
-        self.maximum_a_posteriori = maximum_a_posteriori
-        self.log_evidence_mean = log_evidence_mean
-        self.log_evidence_variance = log_evidence_variance
